sA/requirements.txt
Flask==2.0.3
pyjwt==2.1.0
requests==2.26.0
werkzeug==2.2.3
loguru==0.7.2
prometheus_client==0.21.0
prometheus_flask_exporter==0.23.1